move worker into separate file
